requirements.txt
GitPython
plottly